# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CIVORA API
  description: |
    CIVORA API - open digital infrastructure for public-interest service delivery.

    Version: 0.2.0-service-delivery

    ## Public-Interest Service Delivery Lifecycle

    Full end-to-end workflow for public service delivery:
    Request → Person registered → Case created (NEW) → OPEN →
    Eligibility assessed → Evidence collected → IN_REVIEW →
    Assessment completed → DECISION_PENDING → Decision made →
    APPROVED/REJECTED → IN_PROGRESS → Assistance provided →
    FOLLOW_UP → CLOSED

    State machine:
    NEW → OPEN → IN_REVIEW → ASSESSMENT → DECISION_PENDING →
    APPROVED → IN_PROGRESS → FOLLOW_UP → CLOSED
    REJECTED → CLOSED
  version: "0.2.0"
  contact:
    name: CIVORA
    url: https://github.com/alrazihi/civora
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server. Override with CIVORA_API_URL in production.
security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. The token must include `organization_id` claim.
        The API validates that the token's organization matches the `{orgId}`
        path parameter on all tenant-scoped endpoints.
  headers:
    X-Request-ID:
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      description: Correlation ID for the request. Generated if not provided by the client.
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: "title is required"
...
    Case:
      type: object
      required: [id, organization_id, case_number, title, status, service_type, priority, created_by, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        organization_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        case_number:
          type: string
          example: "CAS-20260906-00001234"
        person_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [NEW, OPEN, IN_REVIEW, ASSESSMENT, DECISION_PENDING, APPROVED, REJECTED, IN_PROGRESS, FOLLOW_UP, CLOSED]
        service_type:
          type: string
          enum: [GENERAL, EMERGENCY, FINANCIAL, FOOD, SHELTER, MEDICAL, EDUCATION, TRANSPORT]
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
        created_by:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        assigned_to:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
    CreateCaseRequest:
      type: object
      required: [title, service_type, priority]
      properties:
        title:
          type: string
          maxLength: 300
        description:
          type: string
          maxLength: 5000
        service_type:
          type: string
          enum: [GENERAL, EMERGENCY, FINANCIAL, FOOD, SHELTER, MEDICAL, EDUCATION, TRANSPORT]
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
        person_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
    ChangeCaseStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [NEW, OPEN, IN_REVIEW, ASSESSMENT, DECISION_PENDING, APPROVED, REJECTED, IN_PROGRESS, FOLLOW_UP, CLOSED]
    AssignCaseRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
...
    AuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        organization_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        actor_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
          nullable: true
        action:
          type: string
        resource:
          type: string
        resource_id:
          type: string
          nullable: true
        outcome:
          type: string
          enum: [success, failure]
        request_id:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        hash:
          type: string
        previous_hash:
          type: string
          nullable: true
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /organizations/{orgId}/cases:
    post:
      operationId: createOrgcases
      tags: [Cases]
      summary: Create a case
      description: Creates a new case (service request) within the organization. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: Idempotency-Key
          in: header
          description: |
            Idempotency key for safe retries. Guarantees at-most-once execution
            for the same key within the same organization.
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        '201':
          description: "Case created (status: CREATED)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      operationId: getOrgcases
      tags: [Cases]
      summary: List cases
      description: Lists cases in the organization. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [NEW, OPEN, IN_REVIEW, ASSESSMENT, DECISION_PENDING, APPROVED, REJECTED, IN_PROGRESS, FOLLOW_UP, CLOSED]
          description: Filter cases by status
      responses:
        '200':
          description: List of cases with pagination
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Case'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}/cases/{caseId}:
    get:
      operationId: getOrgcasesCase
      tags: [Cases]
      summary: Get a case
      description: Retrieves a specific case. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      responses:
        '200':
          description: Case details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/cases/{caseId}/transitions:
    post:
      operationId: createOrgcasesCasetransitions
      tags: [Cases]
      summary: Change case status
      description: |
        Transitions a case to a new status. Invalid transitions are rejected.

        Valid transitions:
        - NEW → OPEN
        - NEW → IN_REVIEW
        - OPEN → IN_REVIEW
        - IN_REVIEW → ASSESSMENT
        - IN_REVIEW → OPEN (reopen)
        - ASSESSMENT → DECISION_PENDING
        - DECISION_PENDING → APPROVED
        - DECISION_PENDING → REJECTED
        - APPROVED → IN_PROGRESS
        - REJECTED → CLOSED
        - IN_PROGRESS → FOLLOW_UP
        - FOLLOW_UP → CLOSED
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeCaseStatusRequest'
      responses:
        '200':
          description: Case with updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /organizations/{orgId}/cases/{caseId}/assign:
    post:
      operationId: createOrgcasesCaseassign
      tags: [Cases]
      summary: Assign case to a user
      description: Assigns a case to a user. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCaseRequest'
      responses:
        '200':
          description: Case with assignment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/audit:
    get:
      operationId: getOrgaudit
      tags: [Audit]
      summary: List audit events
      description: Lists audit events for the organization. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: List of audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
...